Started work on operator extension. Moved sample queries.
