+++ project/Projects.scala
@@ -9,4 +9,5 @@ object Projects {
   lazy val cpgvalidator = project.in(file("cpgvalidator"))
   lazy val cpg2overflowdb = project.in(file("cpg2overflowdb"))
   lazy val console = project.in(file("console"))
+  lazy val queries = project.in(file("queries"))
 }
