+++ queries/build.sbt
@@ -1,0 +1,9 @@
+name := "queries"
+
+dependsOn(Projects.semanticcpg % "compile -> compile; test -> test",
+          Projects.dataflowengine % "compile -> compile; test -> test")
+
+libraryDependencies ++= Seq(
+  "org.scalatest" %% "scalatest" % "3.0.8" % Test,
+  "io.shiftleft" %% "fuzzyc2cpg" % Versions.fuzzyc2cpg % Test
+)
