+++ queries/src/test/scala/io/shiftleft/queries/CopyOperations.scala
@@ -1,9 +1,9 @@
-package io.shiftleft.semanticcpg.samplequeries
+package io.shiftleft.queries
 
 import io.shiftleft.codepropertygraph.Cpg
+import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.testfixtures.CodeToCpgFixture
 import org.scalatest.{Matchers, WordSpec}
-import io.shiftleft.semanticcpg.language._
 
 class CopyOperations extends WordSpec with Matchers {
 
@@ -26,9 +26,7 @@ class CopyOperations extends WordSpec with Matchers {
       * access. For that access, determine the destination buffer and all indices
       * */
     def indexBufferAssigns(cpg: Cpg) =
-      cpg
-        .call(".*assign.*")
-        .argument(1)
+      cpg.assign.target
         .ast
         .isCall
         .name(".*op.*computedMemberAccess.*")
