+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/operatorextension/NodeTypeStarters.scala
@@ -1,0 +1,10 @@
+package io.shiftleft.semanticcpg.language.operatorextension
+
+import io.shiftleft.codepropertygraph.Cpg
+import io.shiftleft.semanticcpg.language._
+
+class NodeTypeStarters(cpg: Cpg) {
+
+  def assign: Assignment = new Assignment(cpg.call.name("<operator>.assignment.*").raw)
+
+}
